move sampler to otel package
